refactor: convert users creation from modal to standalone page
