+++ apps/web/app/modules/Admin/Users/User.page.tsx
@@ -1,8 +1,7 @@
 import { useParams } from "@remix-run/react";
 import { startCase } from "lodash-es";
-import { useState } from "react";
 import { useForm } from "react-hook-form";
-import { UpdateUserBody } from "~/api/generated-api";
+import type { UpdateUserBody } from "~/api/generated-api";
 import { useAdminUpdateUser } from "~/api/mutations/admin/useAdminUpdateUser";
 import { userQueryOptions, useUserById } from "~/api/queries/admin/useUserById";
 import { Button } from "~/components/ui/button";
@@ -25,7 +24,6 @@ const User = () => {
 
   const { data: user, isLoading } = useUserById(id);
   const { mutateAsync: updateUser } = useAdminUpdateUser();
-  const [isEditing, setIsEditing] = useState(false);
 
   const {
     control,
@@ -33,46 +31,32 @@ const User = () => {
     formState: { isDirty },
   } = useForm<UpdateUserBody>();
 
-  if (isLoading)
+  if (isLoading) {
     return (
       <div className="flex items-center justify-center h-full">
         <Loader />
       </div>
     );
+  }
+
   if (!user) throw new Error("User not found");
 
   const onSubmit = (data: UpdateUserBody) => {
     updateUser({ data, userId: id }).then(() => {
       queryClient.invalidateQueries(userQueryOptions(id));
-      setIsEditing(false);
     });
-    setIsEditing(false);
   };
 
   return (
-    <div className="bg-white rounded-lg shadow-md p-6 max-w-4xl mx-auto">
-      <form
-        onSubmit={handleSubmit(onSubmit)}
-        className="bg-white rounded-lg h-full p-4"
-      >
+    <div className="flex flex-col">
+      <form onSubmit={handleSubmit(onSubmit)} className="rounded-lg h-full">
         <div className="flex items-center justify-between">
           <h2 className="text-2xl text-neutral-950 font-semibold mb-4">
             User Information
           </h2>
-          {isEditing ? (
-            <div>
-              <Button type="submit" disabled={!isDirty} className="mr-2">
-                Save
-              </Button>
-              <Button type="button" onClick={() => setIsEditing(false)}>
-                Cancel
-              </Button>
-            </div>
-          ) : (
-            <Button type="button" onClick={() => setIsEditing(true)}>
-              Edit
-            </Button>
-          )}
+          <Button type="submit" disabled={!isDirty} className="mr-2">
+            Save
+          </Button>
         </div>
         <div className="space-y-4 pt-4">
           {displayedFields.map((field) => (
@@ -80,12 +64,7 @@ const User = () => {
               <Label className="text-neutral-600 font-normal">
                 {field === "archived" ? "Status" : startCase(field)}
               </Label>
-              <UserInfo
-                name={field}
-                control={control}
-                isEditing={isEditing}
-                user={user}
-              />
+              <UserInfo name={field} control={control} isEditing user={user} />
             </div>
           ))}
         </div>
